Ensure pi matches dtype

+++ train_utils.py
@@ -28,7 +28,7 @@ def gauss_l2(x_mu, x_sigma, x):
     x_var = x_sigma ** 2
     x_log_var = x_var.log()
     squared_diff_normed = torch.true_divide(squared_difference, x_var)
-    return 0.5 * (torch.log(2 * torch.tensor(math.pi, device=x.device)) + x_log_var + squared_diff_normed).mean(-1)
+    return 0.5 * (torch.log(2 * torch.tensor(math.pi, dtype=x.dtype, device=x.device)) + x_log_var + squared_diff_normed).mean(-1)
 
 
 ### training stuff ###
